cart-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
python-jose==3.3.0
python-dotenv==1.0.0
pydantic==2.4.2 